packages
.

+++ domain/solution.py
@@ -8,15 +8,15 @@ import subprocess
 import psutil
 import sys
 class Solution:
-    _MAX_RAM_USAGE =50 #in percentage
+    _MAX_RAM_USAGE =30 #in percentage
     _DIRECTIVES_FILENAME = 'directives.tcl'
     _VIVADO_PROCESSNAME = 'vivado_hls'
     _SCRIPT_PATH = './callVivado.sh'
     _FF_VALUE = 1; _LUT_VALUE = 2; _DSP_VALUE = 345.68; _BRAM_VALUE = 547.33
     
     
-    def __init__(self,diretivas:dict, cFile, prjFile):
-        self.diretivas = copy.deepcopy(diretivas)
+    def __init__(self,directives:dict, cFile, prjFile):
+        self.directives = copy.deepcopy(directives)
         self.cFile = cFile
         self.prjFile = prjFile
         if sys.platform == 'win32':
@@ -35,7 +35,7 @@ class Solution:
         """
         set directives used in this solution
         """
-        self.diretivas = copy.deepcopy(directives)
+        self.directives = copy.deepcopy(directives)
 
     def setOneResult(self,key,value):
         for resultKey in self.resultados:
@@ -50,13 +50,17 @@ class Solution:
                 
     def __writeDirectivesIntoFile(self):
         directivesFile = open(self._DIRECTIVES_FILENAME, "w")
-        for value in self.diretivas.values():
+        for value in self.directives.values():
             if value != '' and value is not None:
                 directivesFile.write(value + '\n')
             print(value)
         directivesFile.close()  
 
-    def runSynthesisTeste(self):
+    def runSynthesisTeste(self,timeLimit=None):
+        if timeLimit is None:
+            timeLimit = float('inf')
+        if timeLimit<=0:
+            raise Exception("****Vivado_HLS has exceed max time usage****")
         resultados = {}
         resultados['FF'] = randrange(100)
         resultados['DSP'] = randrange(100)
@@ -67,14 +71,27 @@ class Solution:
 
         self.resultados = resultados
    
-    def runSynthesis(self):
+    def runSynthesis(self,timeLimit = None):
         #TODO antes de chamar a sintese verificar se ja tem um processo do vivado rodando e fazer esse processo n ser confundido com o que vamos rodar
         #path to synthesis data
         xml='./Raise_dse/solution1/syn/report/csynth.xml'
 
+        #if not especified, there is infinite time to run synthesis
+        if timeLimit is None:
+            timeLimit = float('inf')
+        if timeLimit<=0:
+            raise Exception("****Vivado_HLS has exceed max time usage****")
+            
         mydir='./Raise_dse'
         if os.path.exists(mydir):
-            shutil.rmtree(mydir)
+            try:
+                shutil.rmtree(mydir)
+            except Exception as error:
+                for proc in psutil.process_iter(['name']):
+                    if proc.name() == self._VIVADO_PROCESSNAME:
+                        proc.kill()
+                        break
+                shutil.rmtree(mydir)
         self.__writeDirectivesIntoFile()
         print('Running Synthesis...')
         #vivado call using subprocess
@@ -83,9 +100,12 @@ class Solution:
         vivadoIsRunning = True
                     
         time.sleep(2) #para dar tempo de iniciar vivado
+        start = time.time()
+        
+
         while vivadoIsRunning:
             #tempo entre duas checagens de se o vivado continua rodando
-            time.sleep(1)
+            time.sleep(3)
             vivadoIsRunning = False
             for proc in psutil.process_iter(['name']):
                 if proc.name() == self._VIVADO_PROCESSNAME:
@@ -93,11 +113,16 @@ class Solution:
                     try:
                         memoryUse = proc.memory_percent()
                     except Exception as e:
+                        #proc.terminate()
                         print(e)
                         break
                     if memoryUse > self._MAX_RAM_USAGE:
-                        proc.terminate()
+                        proc.kill()   
                         raise Exception("****Vivado_HLS has exceed max RAM usage****")
+                    
+                    if time.time()-start >= timeLimit:
+                        proc.kill()   
+                        raise Exception("****Vivado_HLS has exceed max time usage****")
                     break
             
         if os.path.exists(xml):
